Deklaration der JavaFX-Anwendung aus Main- in separate Klasse extrahiert, um TrayIcon auch unter Mac OS X zu ermöglichen
